% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earthcircle.R
\name{earthcircle}
\alias{earthcircle}
\title{Generate coordinates of an "earth" circle.}
\usage{
earthcircle(x, y = NULL, scale = 1852 * 60, ..., n = 36, from = 0, to = 2 * pi)
}
\arguments{
\item{x}{longitude of central location for circle, or lon,lat together in matrix, data frame, or list}

\item{y}{latitude of location (ignored if 'x' includes y)}

\item{scale}{the scale of the circle, large enough default to see on world maps}

\item{...}{ignored currently}

\item{n}{the number of coordinates to provide each circle}

\item{from}{the minimum radial angle (default 0)}

\item{to}{the maximum radial angle}
}
\value{

}
\description{
A circle on the ground for every input longitude,latitude.
}
\examples{
x <- earthcircle(cbind(c(0, -50), c(0, -90)))
plot(earthcircle:::.prj(x, "+proj=laea +lat_0=-90"), asp = 1, type = "l")
}
